// x-pack/solutions/observability/plugins/observability/public/components/alert_search_bar/types.d.ts
import type { ReactElement } from 'react';
import type { FilterGroupHandler } from '@kbn/alerts-ui-shared';
import { type FilterControlConfig } from '@kbn/alerts-ui-shared';
import type { HttpStart } from '@kbn/core-http-browser';
import type { ToastsStart } from '@kbn/core-notifications-browser';
import { type NotificationsStart } from '@kbn/core-notifications-browser';
import type { DataPublicPluginStart, TimefilterContract } from '@kbn/data-plugin/public';
import { type SavedQuery } from '@kbn/data-plugin/public';
import type { DataViewsPublicPluginStart } from '@kbn/data-views-plugin/public';
import type { SpacesPluginStart } from '@kbn/spaces-plugin/public';
import type { AlertsSearchBarProps } from '@kbn/triggers-actions-ui-plugin/public/application/sections/alerts_search_bar';
import type { BoolQuery, Filter } from '@kbn/es-query';
import type { IUiSettingsClient } from '@kbn/core-ui-settings-browser';
import type { AlertStatus } from '../../../common/typings';
export interface AlertStatusFilterProps {
    status: AlertStatus;
    onChange: (id: AlertStatus) => void;
}
export interface AlertSearchBarWithUrlSyncProps extends CommonAlertSearchBarProps {
    urlStorageKey: string;
    defaultState?: AlertSearchBarContainerState;
    disableLocalStorageSync?: boolean;
}
export interface Dependencies {
    data: {
        query: {
            timefilter: {
                timefilter: TimefilterContract;
            };
        };
    };
    triggersActionsUi: {
        getAlertsSearchBar: (props: AlertsSearchBarProps) => ReactElement<AlertsSearchBarProps>;
    };
    useToasts: () => ToastsStart;
}
export interface Services {
    timeFilterService: TimefilterContract;
    AlertsSearchBar: (props: AlertsSearchBarProps) => ReactElement<AlertsSearchBarProps>;
    http: HttpStart;
    data: DataPublicPluginStart;
    dataViews: DataViewsPublicPluginStart;
    notifications: NotificationsStart;
    spaces?: SpacesPluginStart;
    useToasts: () => ToastsStart;
    uiSettings: IUiSettingsClient;
}
export interface ObservabilityAlertSearchBarProps extends AlertSearchBarContainerState, AlertSearchBarStateTransitions, CommonAlertSearchBarProps {
    services: Services;
    filterControls?: Filter[];
    onFilterControlsChange: (filterControls: Filter[]) => void;
    savedQuery?: SavedQuery;
    showFilterBar?: boolean;
    disableLocalStorageSync?: boolean;
}
export interface AlertSearchBarContainerState {
    rangeFrom: string;
    rangeTo: string;
    kuery: string;
    status?: AlertStatus;
    filters?: Filter[];
    savedQueryId?: string;
    controlConfigs?: FilterControlConfig[];
    groupings?: string[];
}
interface AlertSearchBarStateTransitions {
    onRangeFromChange: (rangeFrom: string) => void;
    onRangeToChange: (rangeTo: string) => void;
    onKueryChange: (kuery: string) => void;
    onStatusChange?: (status: AlertStatus) => void;
    onFiltersChange?: (filters: Filter[]) => void;
    setSavedQuery?: (savedQueryId?: SavedQuery) => void;
    onControlConfigsChange?: (controlConfigs: FilterControlConfig[]) => void;
}
interface CommonAlertSearchBarProps {
    appName: string;
    onEsQueryChange: (query: {
        bool: BoolQuery;
    }) => void;
    defaultFilters?: Filter[];
    filterControls?: Filter[];
    onFilterControlsChange: (filterControls: Filter[]) => void;
    onControlApiAvailable?: (controlGroupHandler: FilterGroupHandler | undefined) => void;
}
export {};